refactor: internal/registrar pkg
for easier file handling
