+++ .env.example
@@ -26,6 +26,7 @@ LOG_DIR=data/logs/othentic/
 OTHENTIC_CLIENT_RPC_ADDRESS=http://localhost:8545
 MANAGER_IP_ADDRESS=http://localhost:9003
 DATABASE_IP_ADDRESS=http://localhost:9002
+DATBASE_DOCKER_IP_ADDRESS=localhost
 AGGREGATOR_IP_ADDRESS=127.0.0.1
 HEALTH_IP_ADDRESS=http://localhost:9004
