+++ cmd/registrar/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"fmt"
-	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 
@@ -14,31 +13,24 @@ import (
 var logger logging.Logger
 
 func main() {
-
 	if err := logging.InitLogger(logging.Development, logging.RegistrarProcess); err != nil {
 		panic(fmt.Sprintf("Failed to initialize logger: %v", err))
 	}
 	logger = logging.GetLogger(logging.Development, logging.RegistrarProcess)
-	logger.Info("Starting registrar node (poll-based)...")
+	logger.Info("Starting registrar service ...")
 
-	// Initialize registrar configuration
 	config.Init()
+	registrar.InitABI()
 
-	// Get contract addresses
 	avsGovernanceAddress := common.HexToAddress(config.AvsGovernanceAddress)
 	attestationCenterAddress := common.HexToAddress(config.AttestationCenterAddress)
 
-	logger.Info(fmt.Sprintf("Using AVS Governance contract at address: %s", config.AvsGovernanceAddress))
-	logger.Info(fmt.Sprintf("Using Attestation Center contract at address: %s", config.AttestationCenterAddress))
+	logger.Info(fmt.Sprintf("AVS Governance     [L1]: %s", config.AvsGovernanceAddress))
+	logger.Info(fmt.Sprintf("Attestation Center [L2]: %s", config.AttestationCenterAddress))
 
-	// Start the polling service in a goroutine
 	go registrar.StartEventPolling(avsGovernanceAddress, attestationCenterAddress)
 
-	// Keep the program running
-	logger.Info("Registrar node is running.")
+	logger.Info("Registrar service is running.")
 
-	// Keep the main thread alive
-	for {
-		time.Sleep(time.Hour)
-	}
+	select {}
 }
