+++ internal/registrar/fetch.go
@@ -1,0 +1,2 @@
+package registrar
+
